[feat] Create result & result-input page
